Update organisation switcher link logic
Only display the organisation switcher link if a link object is
passed to the component.

This prevents an empty link tag being output to the source if there
is no link destination.

This change still supports the fallback text as long as a link `href`
is supplied.

It also updates the example to show how the component will look
without a change link and with the fallback link text.