Prevent components being initialised twice
Added a catch to all the component I've changed so that they cannot be initialised twice

When a component is initialised, we set a data attribute named for the component. If the attribute is already set, we immediately stop initialising.

I continued the pattern of using jQuery. If we deprecate jQuery in the future we can use elements' dataset property