fix(ticket-panel): Make default ticket panel have no border (#158)
The default ticket panel currently has a black border because `border-left-color` is not specified. This change instead makes it transparent, so that the border only appears if you add one of the variant classes (e.g. `.moj-ticket-panel__content--red`).

This means that default ticket panels have an extra 4px padding on the left, but this doesn't stand out visually and ensures that the text aligns between plain and colourful ticket panels.