fix(add-another): shrink remove button (#256)
The GOV.UK Design System extends buttons to 100% width on small screens. The Add another
component's "Remove" button is absolutely positioned, so it covers up content when it's at 100%
width.

This change shrinks the button back down to `width: auto`, so that it's only as wide as the text
and any content "underneath" can be seen.

I've also added right padding to the title so that it wraps before going underneath the remove
button.

This does reduce the clickable area for the button, which may make it harder for touchscreen users,
however there should not be any other clickable targets nearby, so users are unlikely to press the
wrong thing.