feat(deps): make dependencies explicit
Make govuk-frontend a package dependency, and import its helpers and width-container into SCSS,
where they're used.

Stop building `.govuk-textarea` in SCSS, instead just apply it onto the element directly.

Mark jQuery as a peer dependency, as it's necessary for the JavaScript to work but not bundled as
part of the package.

fixes #147