build: use node lts, sass and npm ci
Build with the LTS version of node, allowing us to use the latest packages. Continue testing on
all current LTSs and v10.

Replace node-sass with sass: node-sass is deprecated and the source of security vulnerabilities.
Continue to use node-sass in tests to ensure builds work.

Use `npm ci` to get precise installs