fix: use inline SVGs
Rather than embedding SVGs as background images, use
them inline. This allows them to keep their parents' fill
colour, so they can always adapt to match the text.

When forced color mode is active (e.g. Windows High
Contrast mode), we need to force the fill colour to match
the context (i.e. text or link colour), otherwise it doesn't
pass through properly on Windows.

This is an accessibility enhancement as it means that
users who use tools like high-contrast mode to force
changes in background and foreground colours will
continue to be able to see iconography. Without it, icons
are invisible if they're the same colour as the forced
background colour.

Fixes #162