feat(prototype kit integration): users of prototype-kit-v13 should have macros and filters available (#378)
In version 13 of GOV.UK Prototype Kit we have some new integration points.  This makes use of the
new integration points and makes the installation simpler for users.  Specifically, for
moj-frontend, this will import all the nunjucks macros (components) and filters without the user
having to do any setup themselves.