fix(scrollable): fix CSS shadow effect
The `.moj-scrollable-pane` class is supposed to add shadows to indicate that its contents are
scrollable, but it is not suitably dynamic.

This CSS uses [scroll shadows][scroll-shadows] to make the shadows dynamically appear on the left
and right when there's overflow content there.

If there's no overflow content, or the container is scrolled all the way to one side, the shadows
are hidden.

I removed the `.moj-scrollable-pane__wrapper` class because it's unnecessary.

[scroll-shadows]: https://www.bram.us/2019/10/24/pure-css-scroll-shadows-vertical-horizontal/