Fix sortable tables for row headers
When the first cell in a row is a header (i.e using the
`firstCellIsHeader` option in the `govukTable` Nunjucks macro), the
index fetched for the column is incorrect (e.g clicking on the second
column sorts the third column).

This updates the finder in the `sort` function, which looks for a table
cell OR a table header within the rows to ensure the index is correct.