fix(prototype): better support for GOV.UK Prototype Kit
Allow the Pattern Library to automatically be available in the Prototype Kit when installed.

This requires producing a file with jQuery bundled (`all.jquery.min.js`) and an init file
(`init.js`). When loaded into the Prototype Kit they automatically initialise all the components.

Update the documentation to explain that no further steps are necessary.

Fixes #465