docs: add masthead to home page
Create a specific "home" layout so that we can insert content above main and
after the page title. Move
all HTML into here to so that the `index.md` file only contains markdown/prose
content.

Reuse the `app-masthead` class, redesigning it to match the new style and add a
background image in MOJ style.