build(eleventy): amend 11ty config to allow reloading components from src directory
Documentation examples by default load components from the built /package directory.  This PR adds
the ability to pass an ENV='dev' flag to the eleventy command to enable loading of components direct
from the /src directory. This also adds a gulp watcher to rebuild the package css and js files
during development.

This needs further work! We have three file watchers running in parallel to build/rebuild the
docs site in development. (11ty, webpack, and gulp) It would be good if we could use just one!