refactor(date picker): add calendar button via JS for pregressive enhancement
This PR moves the calendar popup toggle from in the template to within the JS, as the button
shouldn't show up unless JS is available to toggle the popup dialog.  Improvements were made to the
accessible labelling of the table headers and the individual day buttons to aid screenreader users.