feat(date picker): allow component to be configured via JS as well as via data-attributes
This change updates the component to allow for component properties such as minDate, maxDate and
excludedDates/Days to be passed in via the JS config object. This follows the GDS convention where
component defaults are overridden by the JS config, which is overridden by data attributes. The code
to do this is largely borrowed from GOV.UK frontend.  Also added in this commit are JSDoc comments
for all functions with arguments.