refactor(date picker): remove need for duplicate date regex
The regex for a date was duplicated in the code.  It wasn't necessary to wrap getting the date in
the input in a conditional due to the fact that fomattedDateFromString falls back to todays date
meaning if there is no date in the input or the input contains an invalid date, it will fall back to
setting the currentDate for the calendar to today, which is what is required.