build(dockerfile): update dockerfile to handle deploying to production
Amend dockerfile to be multistage and allow generating a different image for staging and production.
Production needs to be different as it needs to not have http auth, not have the staging banner, and
it needs to be built from a git repo to eanble the 'last updated' functionality for components.