Changing pagination results number (#956)
* docs(pagination): update results indexes on pagination example

The paginion component example shows as page 2, but also outputs
"showing results 10-20 of 30". This commit updates that to be results
11-20 as would actually be displayed.

* fix(pagination): fix aria label for pages to use total pages instead of total results

The `aria-label` for each page entry was of the form "page x of y" where y was the total number of
results, rather than the total number of pages. This PR fixes the error

---------

Co-authored-by: Chris Pymm <chris@chrispymm.co.uk>