docs(examples): add dynamic resizing to ensure iframe is always the correct size
On smaller screens the fixed height of the iframes often meant a lot of wasted whitespace.  This
commit adds and IFrameResizer class to dynamically size/resize the iframes according to the
contents.  This will dynamically update when the iframe content changes height using a
ResizeObserver. It will also resize when elements within the iframe change using a MutationObserver,
such as when the datepicker or buton menu pops up.