docs(example template): amending example template (#979)
* docs(example template): amending example template

Related to https://github.com/ministryofjustice/moj-frontend/issues/958

* docs(examples): fix margin and padding around example tabs

* docs(examples): add dynamic resizing to ensure iframe is always the correct size

On smaller screens the fixed height of the iframes often meant a lot of wasted whitespace.  This
commit adds and IFrameResizer class to dynamically size/resize the iframes according to the
contents.  This will dynamically update when the iframe content changes height using a
ResizeObserver. It will also resize when elements within the iframe change using a MutationObserver,
such as when the datepicker or buton menu pops up.

* docs(examples): adjust scope of the mutationObserver to ensure it triggers size adjustment

* docs(example template): remove commented out line

* docs(example template): fix typo

---------

Co-authored-by: Chris Pymm <chris@chrispymm.co.uk>