docs(examples): update example css to reflect real-life usage
The code in our examples was using application.css, which is a sass compiled file including all of
govuk frontend, moj-frontend and all the custom styles for the docs site.  This could cause
potential issues where our own css could affect components. This commit refactors things such that
the example layout includes compiled govuk-frontend and (separately) compiled moj-frontend and none
of the docs site css. This ensure we are presenting (and testing) components in a more isolated way
with no danger of style leakage.