fix(interruption card): fix specificity conflicts from govuk-frontend (#1540)
* docs: compile a local version of govuk frontend for docs

* fix(interruption card): remove govuk classes to resolve scoping/specidicuty issues

There had been reports of both the heading and the body text appearing black.

This was likley due to the .govuk-heading-l and .govuk-body-l classes
having the same specificity as the .moj-interruption-card__heading and
.moj-interruption-card__body classes.

To resolve this, the .govuk- classes have been removed, and their styles
directly included in the .moj- classes. This means .govuk styles cannot
override the component syles.  This also has the added advantage that
the text sizing is now baked-in to the component, and users will not be
tempted to change the .govuk-body-l class for another size class.