fix(date picker): fix enter key not submitting containing form (#1594)
fix #1591

Fixes an issue where presing Enter anywhere in a form containing a datepicker would not submit the form as it should.

Some of the buttons within the datepicker did not have type="button" so were being activated by the Enter press as they were the first "submit" button in the form.