docs: update example shortcode to automatically include arguments
Previously each example njk file needed to include a refernce to the the arguments file in its
frontmatter.  This was a bit redundant as the arguments file always has the same name as the
component and is always in the same location.  This PR updates the shortcode to extract the
arguments file name from the calling page, and automatically include the arguments markup if the
file exists.