fix(datepicker): fix mindate issues caused by BST timezone
This PR fixes inconsistencies in `minDate` / `excludedDate` behaviour due to the time difference caused
by BST. The result of this was that `minDate` was sometimes inclusive
and sometimes exclusive dependent on the timezone.

BREAKING CHANGE: `minDate` is now always inclusive. This means that the date passed for `minDate` will
be available for selection.

fix #923