fix(pagination): fix page count totals in aria labels (#1904)
Previously the page count tataks used `items.length` to calculate the last page number.  If the
items array contains 'dots' elemnts the length will be incorrect.  

This PR adjusts to use the text (number) of the last item in the array as the total pages count which should always be correct.