fix minDate bug (#1884) (#1887)
* fix(datepicker): fix mindate issues caused by BST timezone

This PR fixes inconsistencies in `minDate` and `excludedDate` behaviour due to the time difference caused
by BST. The result of this was that `minDate` was sometimes inclusive and sometimes exclusive dependent on the timezone.

BREAKING CHANGE: `minDate` is now always inclusive. This means that the date passed for `minDate` will
always be available for selection.  You will need to amend the dates you pass to the component if you had been relying on the `minDate` not being selectable.

fix #923

* test: add tests for minDates with and without leading zeros

* fix minDate bug (#1884)

* fix(date picker): ensure minDate is always exclusive and is selectable

---------

Co-authored-by: bill-moj <167539377+bill-moj@users.noreply.github.com>