fix(multifileupload): amend calls to callback functions to use statusText instead of responseText (#1922) (#1932)
Currently the calls to the callback functions use `xhr.responseText` which throws an error due to the
`responseType` being 'json'.  This error was introduced when migrating from jQuery which always has
`responseText` available.  Looking at the type definitions and the old jQuery implementation it appears that this argument was intended to be `statusText`.