feat(scrollable pane): force requirement for region tabindex and label attributes
In order to ensure accessibility, the scrollable nature of the table won't be enabled until all the
required attributes are configured on the table.

BREAKING CHANGE: In order for the content in the pane to scroll, the component MUST have
`role="region"`, `tabindex="0"` and either an `aria-labelledby` or `aria-label` attributes. This is
to ensure the component ia accessible.